# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: TeleShop API — crypto marketplace for Telegram Mini Apps
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /shops:
    get:
      operationId: listShops
      tags: [shops]
      summary: List all active shops (marketplace)
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string }
      responses:
        "200":
          description: List of shops
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShopList"
    post:
      operationId: createShop
      tags: [shops]
      summary: Create a new shop
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShopInput"
      responses:
        "201":
          description: Created shop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
        "400":
          description: Validation error
        "409":
          description: Slug already taken
  /shops/me:
    get:
      operationId: getMyShop
      tags: [shops]
      summary: Get current user's shop
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Shop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
        "404":
          description: No shop found
  /shops/{shopId}:
    get:
      operationId: getShop
      tags: [shops]
      summary: Get shop by ID
      parameters:
        - in: path
          name: shopId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Shop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
        "404":
          description: Not found
    patch:
      operationId: updateShop
      tags: [shops]
      summary: Update shop
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: shopId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShopUpdate"
      responses:
        "200":
          description: Updated shop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
...
  /vouchers/validate:
    post:
      operationId: validateVoucher
      tags: [vouchers]
      summary: Validate a voucher code before checkout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoucherValidateInput"
      responses:
        "200":
          description: Voucher details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoucherValidateResult"
        "404":
          description: Invalid or expired voucher
  /vouchers:
    get:
      operationId: listVouchers
      tags: [vouchers]
      summary: List vouchers for merchant's shop
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Voucher list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoucherList"
    post:
      operationId: createVoucher
      tags: [vouchers]
      summary: Create a voucher
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoucherInput"
      responses:
        "201":
          description: Created voucher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Voucher"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      required: [id, telegramId, role, createdAt]
      properties:
        id:
          type: integer
        telegramId:
          type: string
        username:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        role:
          type: string
        referralCode:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    UserUpdate:
      type: object
      properties:
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
    Shop:
      type: object
      required: [id, ownerId, name, slug, plan, status, totalSales, createdAt]
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        plan:
          type: string
        status:
          type: string
        rating:
          type: string
        totalSales:
          type: integer
        walletAddress:
          type: ["string", "null"]
        telegramChannel:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ShopList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Shop"
        total:
          type: integer
    ShopInput:
      type: object
      required: [name, slug, plan]
      properties:
        name:
          type: string
          minLength: 2
        slug:
          type: string
          minLength: 2
        description:
          type: string
        plan:
          type: string
          enum: [starter, growth, pro]
        walletAddress:
          type: string
        telegramChannel:
          type: string
    ShopUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        logoUrl:
          type: string
        bannerUrl:
          type: string
        walletAddress:
          type: string
        telegramChannel:
          type: string
        status:
          type: string
...
    Voucher:
      type: object
      required: [id, code, discountType, discountValue, usedCount, isActive, createdAt]
      properties:
        id:
          type: integer
        shopId:
          type: ["integer", "null"]
        code:
          type: string
        discountType:
          type: string
        discountValue:
          type: string
        minOrderAmount:
          type: ["string", "null"]
        maxUses:
          type: ["integer", "null"]
        usedCount:
          type: integer
        expiresAt:
          type: ["string", "null"]
          format: date-time
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    VoucherList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Voucher"
    VoucherInput:
      type: object
      required: [code, discountType, discountValue]
      properties:
        code:
          type: string
          minLength: 3
        discountType:
          type: string
          enum: [percent, fixed]
        discountValue:
          type: string
        minOrderAmount:
          type: string
        maxUses:
          type: integer
        expiresAt:
          type: string
    VoucherValidateInput:
      type: object
      required: [code, orderAmount]
      properties:
        code:
          type: string
        orderAmount:
          type: string
        shopId:
          type: integer
    VoucherValidateResult:
      type: object
      required: [valid, voucher]
      properties:
        valid:
          type: boolean
        voucher:
          $ref: "#/components/schemas/Voucher"
        discountAmount:
          type: string
        finalAmount:
          type: string
...