# lib/api-zod/src/generated/types/shopUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TeleShop API — crypto marketplace for Telegram Mini Apps
 * OpenAPI spec version: 1.0.0
 */

export interface ShopUpdate {
  name?: string;
  description?: string;
  logoUrl?: string;
  bannerUrl?: string;
  walletAddress?: string;
  telegramChannel?: string;
  status?: string;
}